+++ ctp2_code/gs/outcom/Ic3CivArchive.h
@@ -44,8 +44,8 @@ DEFINE_GUID(CLSID_IC3CivArchive,
 
 DECLARE_INTERFACE_(IC3CivArchive, IUnknown) {
 
-STDMETHOD_ (void, Store) (THIS_ uint8 *pbData, uint32 ulLen) PURE;
-STDMETHOD_ (void, Load) (THIS_ uint8 *pbData, uint32 ulLen) PURE ;
+STDMETHOD_ (void, Store) (THIS_ uint8 *pbData, size_t ulLen) PURE;
+STDMETHOD_ (void, Load) (THIS_ uint8 *pbData, size_t ulLen) PURE ;
 STDMETHOD_ (BOOL, IsStoring)(THIS) PURE ;
 };
 #else
